==> pipeline_1/5. Inference.ipynb <==
import os
import sys
import re
import json
import time
import whisper
import pandas as pd
from tqdm import tqdm
import numpy as np
import datetime
import random
import math
import openai


from youtube_transcript_api import YouTubeTranscriptApi
from get_timestamps import get_timestamps, get_text
from yt_dlp import YoutubeDL
from transformers import pipeline
from datasets import load_dataset, Audio
from datetime import datetime
from googleapiclient.discovery import build

# Url to find timestamps for 
video_url = 'https://www.youtube.com/watch?v=BKpmAGZQetc'


def download_audio(url, output_file):
    # Download audio file from a youtube url using yt-dlp
    ffmpeg_loc = "C:/Users/aladd/Desktop/Shortcuts/ffmpeg/bin/ffmpeg.exe"

    """
    Download audio file from a youtube url using yt-dlp
    """
    ydl_opts = {
        "format": "bestaudio/best",
        "outtmpl": output_file,
        "ffmpeg_location": ffmpeg_loc,
        "postprocessors": [
            {
                "key": "FFmpegExtractAudio",
                "preferredcodec": "wav",
                "preferredquality": "196",
            }
        ],
    }
    with YoutubeDL(ydl_opts) as ydl:
        ydl.download([url])

def convert_srt_to_json(srt):
    """ 
    From YouTubeTranscriptAPI we get a srt file,
    this function converts it to a json file and 
    formats it in the same structure as Whisper output.
    """
    json_list = []

    for dict in srt:
        start_time = dict["start"]
        end_time = dict["start"] + dict["duration"]
        json_list.append(
            {
                "start": start_time,
                "end": end_time,
                "text": dict["text"],
            }
        )
    
    result = {"segments":json_list}
    return result
    
def download_transcript_from_API(video_id):
    """Download transcript from YouTube API"""
    srt = YouTubeTranscriptApi.get_transcript(video_id)
    transcript = convert_srt_to_json(srt)
    return transcript

def download_transcript_using_whisper(video_id, output_file="audio/audio.wav"):
    # make video_id into youtube url
    url = f"https://www.youtube.com/watch?v={video_id}"

    # delete audio.wav if it exists
    if os.path.exists(output_file):
        os.remove(output_file)
        
    # download audio file
    download_audio(url, output_file)

    # transcript using whisper
    # time how long it takes 
    start = time.time()
    result = model.transcribe(output_file, verbose=True)
    end = time.time()
    print(f"Whisper model took {end-start} seconds to transcribe")
    return result

model_name = "facebook/bart-large-cnn"
summarizer = pipeline(
    "summarization",
    model=model_name,
    tokenizer=model_name,
    device="cuda:0",
)

model = whisper.load_model("medium.en")

def summarize_transcript(data):
    subset_texts = []
    text = ""
    start_time = data["segments"][0]["start"]
    idx = 0

    while idx < len(data["segments"]):
        segment = data["segments"][idx]
        segment_start_time = segment["start"]
        segment_end_time = segment["end"]
        segment_text = segment["text"]

        # add segment_text to text
        text += " " + segment_text.strip() if text != "" else segment_text.strip()

        # check if text is longer
        if len(text) > 1500:
            subset_texts.append([start_time, segment_end_time, text])
            start_time = segment_end_time
            text = ""
            idx -= 5
        idx += 1
    
    # add the last text
    subset_texts.append([start_time, segment_end_time, text])

    #print("Starting summarization")
    summary_dict = summarizer(
        [subset[2] for subset in subset_texts],
        max_length=100,
        min_length=0,
    )

    return summary_dict, subset_texts

video_id = video_url.split("=")[1]

try:    
    result = download_transcript_from_API(video_id)
except:
    print("Failed to download transcript from API, using Whisper instead")
    result = download_transcript_using_whisper(video_id)
summary_dict, subset_texts = summarize_transcript(result)

dataset = []
start_end_times = []
overlap_size = 1
summarized_text_string = ""

# take out summary_text from summary_dict
summaries = [summary_d["summary_text"] for summary_d in summary_dict]
for subset, summary_d in zip(subset_texts, summary_dict):
    start_end_times.append([subset[0], subset[1]])

# take out 10 of summary_dict at a time, with overlap of 1
summaries = [summaries[i : i + 10] for i in range(0, len(summaries), 10 - overlap_size)]
start_end_times = [start_end_times[i : i + 10] for i in range(0, len(start_end_times), 10 - overlap_size)]

# for each in summaries join with 1., 2., 3. etc
for i, summary in enumerate(summaries):
    summary = [f"{i+1}. {s}" for i, s in enumerate(summary)]
    summary = "\n".join(summary)
    summary += "\n\n###\n\n"
    dataset.append({"prompt":summary})

ft_model = 'babbage:ft-personal:timestamps-v2-2022-12-10-19-16-54'

keywords = []
openai.api_key = os.environ['API_KEY']

prev_ans = ""
for idx in range(len(dataset)):
    prompt = dataset[idx]["prompt"]
    prompt += f" 1. {prev_ans}\n2" if len(prev_ans) > 0 else ""
    keywords += [prev_ans] if len(prev_ans) > 0 else []
    res = openai.Completion.create(model=ft_model, prompt=prompt, max_tokens=256, temperature=0.0, presence_penalty=0.5, frequency_penalty=0.5, stop=["###"])

    print(prompt + res['choices'][0]['text'])
    print("\n\n")
    
    for each in res['choices'][0]['text'].split("\n"):
        if isinstance(each, str) and len(each) > 0:
            keywords.append(each.split(".")[1].strip())
    
    prev_ans = keywords[-1]

start_end_times = [item for sublist in start_end_times for item in sublist]

# make a unique list of keywords and their start and end times in order
unique_keywords = []
for keyword, (start, end) in zip(keywords, start_end_times):
    if len(unique_keywords) == 0:
        unique_keywords.append([keyword, [start, end]])
    elif keyword != unique_keywords[-1][0]:
        unique_keywords.append([keyword, [start, end]])
    else:
        unique_keywords[-1][1][1] = end

print(unique_keywords)




classifier = pipeline("zero-shot-classification",
                      model="facebook/bart-large-mnli")


# iterate through every segment and print start, end, text
time_stamps = []
print(len(result["segments"]))
for segment in tqdm(result["segments"]):
    sequence_to_classify = segment['text']
    start_seg = segment['start']
    end_seg = segment['end']
    # take out 3 keywords from unique_keywords that are closest to start_time and end_time
    i = np.where([start_seg>=start_key for keyword, (start_key, end_key) in unique_keywords])[0].max()
    # add i-1, i, i+1 to candidate_labels if they exist
    candidate_labels = []
    if i-1 >= 0:
        candidate_labels.append(unique_keywords[i-1][0])
    candidate_labels.append(unique_keywords[i][0])
    if i+1 < len(unique_keywords):
        candidate_labels.append(unique_keywords[i+1][0])
    res = classifier(sequence_to_classify, candidate_labels)
    # get max label and add to time_stamps
    max_label = res['labels'][res['scores'].index(max(res['scores']))]
    time_stamps.append([max_label, start_seg])

time_stamps[0][1] = 0

for each in time_stamps:
    if type(each[1]) == str:
        # convert hh:mm:ss to seconds
        each[1] = sum(x * int(t) for x, t in zip([3600, 60, 1], each[1].split(":")))

# iterate through time_stamps and only keep the first instance of each label each time it occurs
# iterate through time_stamps and print
def clean_time_stamps(time_stampo):
    time_stamps = time_stampo.copy()
    unique_time_stamps = []
    for time_stamp in time_stamps:
        if len(unique_time_stamps) == 0:
            unique_time_stamps.append(time_stamp)
            
        elif time_stamp[0] != unique_time_stamps[-1][0]:
            unique_time_stamps.append(time_stamp)
    
     # go through unique_time_stamps and remove any that are within 30 seconds of the next one, remove repeated labels 
    i = 0
    while i < len(unique_time_stamps)-1:
        if unique_time_stamps[i+1][1] - unique_time_stamps[i][1] < 30:
            unique_time_stamps.pop(i+1)
        elif unique_time_stamps[i+1][0] in [t[0] for t in unique_time_stamps[:i+1]]:
            unique_time_stamps.pop(i+1)
        else:
            i += 1

    # convert unique_time_stamps to h:m:s format and remove ms
    for i, time_stamp in enumerate(unique_time_stamps):
        unique_time_stamps[i][1] = str(datetime.timedelta(seconds=time_stamp[1])).split('.')[0]
    print(f'Second clean: {unique_time_stamps}')

    # print the unique_time_stamps in nice format with time and label
    for time_stamp in unique_time_stamps:
        print(f"{time_stamp[1]}: {time_stamp[0]}")

clean_time_stamps(time_stamps)
